config: resolve schedule dependency
