+++ server/config/util.go
@@ -62,3 +62,16 @@ func ValidateURLWithScheme(rawURL string) error {
 	}
 	return nil
 }
+
+var schedulerMap = make(map[string]struct{})
+
+// RegisterScheduler registers the scheduler type.
+func RegisterScheduler(typ string) {
+	schedulerMap[typ] = struct{}{}
+}
+
+// IsSchedulerRegistered checks if the named scheduler type is registered.
+func IsSchedulerRegistered(name string) bool {
+	_, ok := schedulerMap[name]
+	return ok
+}
