+++ server/schedule/scheduler.go
@@ -21,6 +21,7 @@ import (
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/log"
+	"github.com/tikv/pd/server/config"
 	"github.com/tikv/pd/server/core"
 	"github.com/tikv/pd/server/schedule/operator"
 	"github.com/tikv/pd/server/schedule/opt"
@@ -98,12 +99,7 @@ func RegisterSliceDecoderBuilder(typ string, builder ConfigSliceDecoderBuilder) 
 		log.Fatal("duplicated scheduler", zap.String("type", typ))
 	}
 	schedulerArgsToDecoder[typ] = builder
-}
-
-// IsSchedulerRegistered check where the named scheduler type is registered.
-func IsSchedulerRegistered(name string) bool {
-	_, ok := schedulerMap[name]
-	return ok
+	config.RegisterScheduler(typ)
 }
 
 // CreateScheduler creates a scheduler with registered creator func.
